Refactor API endpoint and improve loading text in StokvelDashboard
- Updated the API endpoint from '/api/my-stokvels' to '/api/stokvels' in StokvelDashboard component.
- Changed loading text from "Loading…" to "Loading...".
- Modified StatusBadge to use a fallback for status from memberships.
